+++ 10xGitHubPolicies.App/Middleware/TestModeAuthenticationHandler.cs
@@ -11,8 +11,8 @@ namespace _10xGitHubPolicies.App.Middleware;
 /// </summary>
 public class TestModeAuthenticationHandler : AuthenticationHandler<AuthenticationSchemeOptions>
 {
-    public TestModeAuthenticationHandler(IOptionsMonitor<AuthenticationSchemeOptions> options, ILoggerFactory logger, UrlEncoder encoder, ISystemClock clock)
-        : base(options, logger, encoder, clock)
+    public TestModeAuthenticationHandler(IOptionsMonitor<AuthenticationSchemeOptions> options, ILoggerFactory logger, UrlEncoder encoder)
+        : base(options, logger, encoder)
     {
     }